% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{A Server Function}
\usage{
server(input, output)
}
\arguments{
\item{input}{shiny input controls of frontend}

\item{output}{shiny and plotly outputs of frontend}
}
\description{
This function allows you manage the logic controls and renders or bindings with statistics functions supported.
}
